Change conversion rules for boolean to floating-point kernels
The convention is now that `Kernel{T}(K)`, where `T` is some element type and
`K` a kernel, always uses the same rules as in Julia for converting element
type.

The new `strel` method is provided to convert a boolean kernel (or a
rectangular box) into a floating-point kernel suitable for greyscale
mathematical morphology (a so-called flat structuring element).